AI-Driven-Code-Review-Developer-Assistant-Tool/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.7
anthropic==0.7.8
tree-sitter==0.20.4
tree-sitter-python==0.23.0
tree-sitter-javascript==0.21.0
pygithub==2.1.1
gitpython==3.1.40
docker==6.1.3
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5
stripe==7.8.0
gunicorn==21.2.0
cryptography==41.0.7
bleach==6.1.0
email-validator==2.1.0
sqlparse==0.4.4
pyopenssl==23.3.0
certifi==2023.11.17
alembic==1.13.0

# RAG and Vector Database Dependencies
chromadb==0.4.18
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.1
numpy==1.24.3
faiss-cpu==1.7.4
langchain==0.0.350
tiktoken==0.5.2
